Refactor item binding and UI in ExplorerBrowser
This commit includes a significant refactoring of the code. The `ItemsSource` for both `TreeView` and `GridView` in `ExplorerBrowser.xaml` has been changed to bind to the same source of items. The `ShellItemCollection` properties and several methods in `ExplorerBrowser.xaml.cs` have been removed and replaced with a new `NavigateTo` method. A commented out `Populate` method has been added to the `IWinUIExplorerBrowser` interface. In `FileManagerViewModel.cs`, a large amount of code has been removed, indicating a change in how item retrieval and updates are handled. In `FileManagerPage.xaml`, several changes have been made to the `CommandBar` elements, suggesting a simplification of the user interface.