Updated ExplorerBrowser UI and functionality
Updated the `ExplorerBrowser.xaml` and `ExplorerBrowser.xaml.cs` files to comment out certain styles and elements, and to replace the `NavigateTo` method with a new `Navigate` method that includes an additional parameter for navigation category. Added a new `GetItemsArray` method to get an array of `ShellItem` objects. The `IWinUIExplorerBrowser.cs` file has been updated to change the `IWinUIExplorerBrowser` interface from internal to public and add a new `ExplorerBrowserNavigationItemCategory` enum. The `FileManagerPage.xaml` file has been updated to rearrange elements and remove the `ContentAreaGrid` element and its children.