Updated ExplorerBrowser UI and added ShellTreeViewItem classes
Significant updates were made to the ExplorerBrowser UI and related classes. The `xmlns:vanarashell` namespace was added to the `ExplorerBrowser.xaml` file, and various templates and styles were updated or uncommented. The `BorderBrush` property of several controls was updated, and the `ItemsSource` property of the `TreeView` control was updated to bind to `m_TreeViewItems`.

In the `ExplorerBrowser.xaml.cs` file, the `m_CurrentFolder` field was updated to use the `ShellFolder` type and the `m_TreeViewItems` field was added. The `GetItemsArray` method was updated to get a list of children from `m_CurrentFolder`.

The `SupportedOSPlatform` attribute was added to the `IWinUIExplorerBrowser` interface. The `ShellTreeViewItem` and `ShellTreeViewItemCollection` classes were added to the `IWinUIExplorerBrowser.cs` file. These classes represent a node in a tree view for a `ShellItem` and a read-only list of `ShellTreeViewItem` objects respectively. A comment was also added to the `NavigationLog` class.